+++ src/entities/book/index.ts
@@ -1,5 +1,5 @@
 export { default as BookCard } from './ui/BookCard';
 export { default as BookList } from './ui/BookList';
-export type { Book } from './model/book';
+export type { Book, BookDetail } from './model/book';
 export { BookListTab } from './model/book-list-tab';
 export { bookQueries } from './api/book-queries';
