+++ src/features/book-note/model/book-note-form-values.ts
@@ -1,0 +1,10 @@
+export interface BookNoteFormValues {
+  readingStatus: string;
+  startDate: Date;
+  endDate: Date;
+  recommended: boolean | null;
+  overallRating: number;
+  content: string;
+  quotes: { text: string; page: string }[];
+  visibility: boolean;
+}
